Adição de classe para tratamento de erros e melhorias nos atributos

+++ api/src/main/java/com/sultsdev/projeto1/domain/dto/DadosCadastroFranquia.java
@@ -1,5 +1,7 @@
 package com.sultsdev.projeto1.domain.dto;
 
+import java.math.BigDecimal;
+
 import com.sultsdev.projeto1.model.Segmento;
 
 import jakarta.validation.constraints.Email;
@@ -15,35 +17,35 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DadosCadastroFranquia {
-	
+
 	@NotBlank
 	private String nome;
-	
+
 	@NotNull
 	private Integer totalUnidades;
-	
+
 	@NotBlank
 	private String estadoSede;
-	
+
 	@NotBlank
 	@Email
 	private String email;
-	
+
 	@NotNull
-	private Double investimentoInicial;
-	
+	private BigDecimal investimentoInicial;
+
 	@NotBlank
 	private String subsegmento;
-	
+
 	@NotBlank
 	private String tipoNegocio;
-	
+
 	@NotBlank
 	private String url;
-	
+
 	@NotNull
 	private Boolean ativo;
-	
+
 	@NotNull
 	private Segmento segmento;
 }
