+++ api/src/main/java/com/sultsdev/projeto1/repository/FranquiaRepository.java
@@ -1,5 +1,6 @@
 package com.sultsdev.projeto1.repository;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import org.springframework.data.domain.Page;
@@ -21,11 +22,10 @@ public interface FranquiaRepository extends JpaRepository<Franquia, Long> {
 			+ "AND (:dataUltimaAtualizacaoStart IS NULL OR f.ultimaAtualizacao >= :dataUltimaAtualizacaoStart) "
 			+ "AND (:dataUltimaAtualizacaoEnd IS NULL OR f.ultimaAtualizacao <= :dataUltimaAtualizacaoEnd) "
 			+ "AND f.ativo = TRUE")
-	Page<Franquia> findAllByFilters(@Param("nome") String nome, 
-			@Param("segmento") String segmento,
-			@Param("estadoSede") String estadoSede, 
-			@Param("investimentoInicialStart") Double investimentoInicialStart,
-			@Param("investimentoInicialEnd") Double investimentoInicialEnd,
+	Page<Franquia> findAllByFilters(@Param("nome") String nome, @Param("segmento") String segmento,
+			@Param("estadoSede") String estadoSede,
+			@Param("investimentoInicialStart") BigDecimal investimentoInicialStart,
+			@Param("investimentoInicialEnd") BigDecimal investimentoInicialEnd,
 			@Param("dataUltimaAtualizacaoStart") LocalDateTime dataUltimaAtualizacaoStart,
 			@Param("dataUltimaAtualizacaoEnd") LocalDateTime dataUltimaAtualizacaoEnd, Pageable pageable);
 }
